requirements.txt
discord.py==2.4.0
opencv-python==4.10.0.84
numpy==2.1.3
mysql-connector-python==9.1.0